Bug fixxing
Vehicle will now slow down when closer to the waypoint, trying to be at rest when capturing it.
A tolerance has been defined to consider that a given WP has been captured.
A bug related to the vehicle's acceleration has been fixed - this bug was making the vehicle change its heading constantly when close to the WP.